pyproject.toml
[tool.poetry]
name = "zeus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
efinance = "^0.5.0"
moomoo-api = "^7.4.3610"
tradingview-ta = "^3.3.0"
sanic = {extras = ["ext"], version = "^23.6.0"}
python-socketio = "^5.9.0"
sanic-ext = "^23.6.0"
mayim = {extras = ["sqlite"], version = "^1.1.0"}
pydantic = "^2.4.2"
mypy-extensions = "^1.0.0"
alpaca-py = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
